Move encryption/encoder to CRDT package
